Changed get wallet.controller body params to query

+++ backend/src/wallet/entities/Transaction-btc.entity.ts
@@ -36,7 +36,7 @@ export class TransactionBTC extends BaseEntity {
   @Column()
   to: string
 
-  @Column('decimal', { precision: 16, scale: 8 })
+  @Column('decimal', { precision: 30, scale: 8 })
   value: number
 
   @ManyToOne(type => WalletBTC, wallet => wallet.transactions, {
